AuthService/src/main.py
from contextlib import asynccontextmanager

import uvicorn
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware

from auth.router import api
from exceptions import (
    BaseHTTPError,
    base_error_handler,
    internal_exception_handler,
)


@asynccontextmanager
async def lifespan(app: FastAPI):
    yield


app = FastAPI(lifespan=lifespan)

app.add_middleware(
    CORSMiddleware,
    allow_origins=[
        "http://localhost:3000",
        "http://127.0.0.1:3000",
        "http://localhost:8080",
        "http://127.0.0.1:8080",
        "http://localhost:8081",
        "http://127.0.0.1:8081",
    ],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

app.include_router(api)
app.add_exception_handler(BaseHTTPError, base_error_handler)
app.add_exception_handler(500, internal_exception_handler)

if __name__ == "__main__":
    uvicorn.run("src.main:app", reload=True)
