AuthService/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["R-Kiselev <sputnavik@gmail.com>"]
readme = "README.md"
packages = [{ include = "auth-service" }]

[tool.poetry.dependencies]
python = ">=3.13"
fastapi = {extras = ["standard"], version = ">=0.115.11,<0.116.0"}
pydantic-settings = ">=2.8.1,<3.0.0"
sqlalchemy = {extras = ["asyncio"], version = ">=2.0.39,<3.0.0"}
asyncpg = ">=0.30.0,<0.31.0"
alembic = "^1.15.1"
bcrypt = "^4.3.0"
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]  # Достаточно poetry-core>=1.0.0
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100
